+++ R/run_sampler.R
@@ -70,6 +70,8 @@ run_sampler_ggm = function(spec) {
   bb_alpha_between = bb_between_or_sentinel(p$beta_bernoulli_alpha_between)
   bb_beta_between = bb_between_or_sentinel(p$beta_bernoulli_beta_between)
 
+  correction = ggm_edge_prior_correction(p, s, d$num_variables)
+
   out_raw = sample_ggm(
     inputFromR = list(
       X = d$x,
@@ -105,7 +107,9 @@ run_sampler_ggm = function(spec) {
     max_tree_depth = s$nuts_max_depth,
     na_impute = m$na_impute,
     missing_index_nullable = m$missing_index,
-    delta = p$delta
+    delta = p$delta,
+    correction_theta = correction$theta,
+    correction_logC = correction$logC
   )
 
   out_raw